% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{calcResid}
\alias{calcResid}
\title{Forward calculation for residual only}
\usage{
calcResid(par, data, cModel, fixParms, fixInis, transParms, sepn, ...)
}
\arguments{
\item{par}{parameters}

\item{data}{data frame}

\item{cModel}{model definition}

\item{fixParms}{fixed parameters}

\item{fixInis}{fixed initial values}

\item{transParms}{tranformation functions}

\item{sepn}{seperater index}

\item{...}{other unused yet inputs}
}
\value{
a vector of residuals
}
\description{
Forward calculation
}

