% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{costSW}
\alias{costSW}
\title{Forward calculation}
\usage{
costSW(data, cModel, optParms, fixParms, optInis, fixInis, transParms = NULL,
  ...)
}
\arguments{
\item{data}{data frame}

\item{cModel}{A lists including the differential equations, the parameters, the states, the initials.}

\item{optParms}{optimazation ODE parameters}

\item{fixParms}{fixed parameters}

\item{optInis}{optimazation initial parameters}

\item{fixInis}{fixed parameters}

\item{transParms}{transformation functions}

\item{...}{observation times}
}
\value{
a data frame
}
\description{
Forward calculation
}

