% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{gen.TOXSWA}
\alias{gen.TOXSWA}
\title{Define the toxswa functions in C}
\usage{
gen.TOXSWA(n = 4, language = c("C", "F95", "Fortran"))
}
\arguments{
\item{n}{number of sediment layers}

\item{language}{the language used}
}
\value{
a list of functions in language
}
\description{
Defind the ODE functions
}

