% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{getY}
\alias{getY}
\title{Get Y}
\usage{
getY(logpar, data, cModel, fixParms, fixInis, transParms, sepn, ...)
}
\arguments{
\item{logpar}{logtransformed parameters}

\item{data}{data frame}

\item{cModel}{model definition}

\item{fixParms}{fixed parameters}

\item{fixInis}{fixed initial values}

\item{transParms}{tranformation functions}

\item{sepn}{seperater index}

\item{...}{other unused yet inputs}
}
\description{
Get Y
}

