% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{simSW}
\alias{simSW}
\title{Simulate a data frame}
\usage{
simSW(func, yini, parms, times, sigmaS, plotting = FALSE)
}
\arguments{
\item{dta}{a sw data frame}
}
\value{
a simulated data frame with
}
\description{
Simu a SW dta
}

