% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ModelA_D.R
\name{SW_Model_calc}
\alias{SW_Model_calc}
\title{Forward calculation}
\usage{
SW_Model_calc(cModel, optParms, fixParms, optInis, fixInis, obsTimes = 1:100,
  transParms = NULL, ...)
}
\arguments{
\item{cModel}{A lists including the differential equations, the parameters, the states, the initials.}

\item{optParms}{optimazation ODE parameters}

\item{fixParms}{fixed parameters}

\item{optInis}{optimazation initial parameters}

\item{fixInis}{fixed parameters}

\item{obsTimes}{observation times}

\item{transParms}{transformation functions}
}
\value{
a data frame
}
\description{
Forward calculation
}

